run from uber jar in cwd

+++ generate-nidmaps.sh
@@ -25,6 +25,17 @@ fi
 mkdir -p "${OUTDIR}"
 echo "OUT:   ${OUTDIR}" 1>&2;
 
-jar=target/yaml2sql-2.4.0-uber.jar
+jar=yaml2sql-2.4.0-uber.jar
+if [ ! -e "${jar}" ]; then
+  if [ ! -e "target/${jar}" ]; then
+    echo "Non existing uber jar" >&2
+    exit 1
+    fi
+  ln -s "target/${jar}"
+  fi
+if [ ! -e "${jar}" ]; then
+  echo "Non existing uber jar" >&2
+  exit 2
+  fi
 java -ea -cp "${jar}" org.oewntk.grind.yaml2sql.Mapper "${IN}" "${IN2}" "${OUTDIR}"
 echo "done $?"
